chore: fix clippy lints

+++ packages/fuels-abigen-macro/tests/harness.rs
@@ -226,7 +226,7 @@ async fn compile_bindings_byte_input() {
     // `SimpleContract` is the name of the contract
     let contract_instance = SimpleContract::new(null_contract_id(), wallet);
 
-    let contract_call = contract_instance.takes_byte(10 as u8);
+    let contract_call = contract_instance.takes_byte(10u8);
 
     let encoded = format!(
         "{}{}",
@@ -535,6 +535,7 @@ async fn create_struct_from_decoded_tokens() {
     );
 
     // Decoded tokens
+    #[allow(clippy::blacklisted_name)]
     let foo = Token::U8(10);
     let bar = Token::Bool(true);
